LPS-66922 Move blogs related data handlers into the correct package
